requirements.txt
smartapi-python
pyotp
pandas
numpy
streamlit